+++ packages/terra-consumer-layout/src/Layout.jsx
@@ -62,7 +62,9 @@ class Layout extends React.Component {
           <a className={cx('skip-to-main-content')} href="#main-container" id="skip-maincontent-link">
             {intl.formatMessage({ id: 'Terra.Consumer.Layout.skipcontent' })}
           </a>
-          {siteAlert}
+          <span>
+            {siteAlert}
+          </span>
         </div>
         <div {...customProps} className={cx('layout', { open: this.state.isMobileNavOpen }, customProps.className)}>
           <nav className={cx('nav')}>